third_party/WebKit/LayoutTests/http/tests/security/script-crossorigin-loads-correctly-credentials-2-expected.txt

CONSOLE ERROR: Access to Script at 'http://localhost:8000/security/resources/cors-script.php?cors=false&credentials=true&value=FAIL' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This test passes if the inserted script fails to load due to CORS.

PASS (expected error reported)
